+++ src/view/app/Indexes/index.tsx
@@ -1,19 +1,18 @@
 import * as React from "react";
 
 import { createRoot } from "react-dom/client";
-import { IConfig } from "../model";
+import { TableDetails } from "../model";
 import "./indexes.css";
 import Indexes from "./indexes";
-
+import { ISettings } from "../../../common/IExtensionSettings";
 
 declare global {
-    interface Window {
-        acquireVsCodeApi(): any;
-        initialData: IConfig;
-    }
-};
-
-const vscode = window.acquireVsCodeApi();
+  interface Window {
+    acquireVsCodeApi(): any;
+    tableDetails: TableDetails;
+    configuration: ISettings;
+  }
+}
 
 const root = createRoot(document.getElementById("root"));
-root.render(<Indexes initialData={window.initialData} vscode={vscode} />);
+root.render(<Indexes tableDetails={window.tableDetails} configuration={window.configuration}/>);