+++ src/view/app/Query/index.tsx
@@ -4,13 +4,14 @@ import { createRoot } from "react-dom/client";
 import "./query.css";
 import QueryForm from "./query";
 import { IOETableData } from "../../../db/oe";
+import { ISettings } from "../../../common/IExtensionSettings";
 
 declare global {
     interface Window {
         acquireVsCodeApi(): any;
         tableData: IOETableData;
         tableName: string;
-        configuration: any;
+        configuration: ISettings;
     }
 }