+++ src/webview/PanelViewProvider.ts
@@ -1,6 +1,6 @@
 import path = require('path');
 import * as vscode from 'vscode';
-import { Constants } from '../common/constants';
+import { Constants } from '../common/Constants';
 import { TableDetails } from '../view/app/model';
 import { TableNode } from '../treeview/TableNode';
 import { TablesListProvider } from '../treeview/TablesListProvider';
@@ -11,6 +11,7 @@ export class PanelViewProvider implements vscode.WebviewViewProvider {
     public _view?: vscode.WebviewView;
     public tableNode?: TableNode;
     public tableListProvider?: TablesListProvider;
+    public readonly configuration = vscode.workspace.getConfiguration("ProBro");
 
     constructor(
         private context: vscode.ExtensionContext, private _type: string) {
@@ -58,7 +59,8 @@ export class PanelViewProvider implements vscode.WebviewViewProvider {
 
         <script>
           window.acquireVsCodeApi = acquireVsCodeApi;
-          window.initialData = ${JSON.stringify(data)};
+          window.tableDetails = ${JSON.stringify(data)};
+          window.configuration = ${JSON.stringify(this.configuration)}
         </script>
     </head>
     <body>