+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 1.2.2 (2023-03-03)
+### Changed
+- Fixed issue with writing socket data
+- Fixed various small bugs
+
+### Added
+- Extension settings configuration for temporary files (-T)
+
+## 1.2.1 (2023-02-17)
+### Changed
+- Fixed critical issue for Linux
+- Fixed various small bugs
+
+### Added
+- Extension settings configuration for logging
+
 ## 1.2.0 (2023-01-30)
 ### Added
 - D file export